Updating NewContractSupplier price  and WeekSupply's sale in dailySchedule

+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/entity/WeekSupply.java
@@ -26,7 +26,7 @@ public class WeekSupply implements BaseEntity{
     private Integer productId;
 
     @Column(name = "price", nullable = false)
-    private Integer price;
+    private Float price;
 
     @Column(name = "sales", nullable = false)
     private Integer sales;
